refactor(engine): move Chromium profile hygiene out of the wwjs adapter
Neither of these two routines is about WhatsApp. They clean up after the
operating system: a process killed with SIGKILL leaves an orphaned browser
behind (Puppeteer's exit hook never runs) and stale SingletonLock/Socket/Cookie
files in the profile dir that block the next launch. They change when Docker,
Puppeteer or the host platform changes — never when the protocol does.

They now live in chromium-profile-hygiene.ts as two free functions taking
(sessionId[, sessionDataPath], logger). Neither read any adapter field beyond
config and the logger, so there is no port back. The adapter keeps the
--openwa-session= marker arg at launch, which is genuinely a launch-arg concern
and is what makes the sweep able to identify its own orphans.

whatsapp-web-js.adapter.spec.ts passes with ZERO edits — 382 tests. Both
describes drive through the public initialize() and spy on shared module
objects, so passing this.logger through keeps the same object under observation.
The child_process import leaves with the block; execFile now appears nowhere in
the adapter.

3281 to 3202 lines. Behaviour identical; both routines were already best-effort
and still log at debug rather than throwing.
